Initial empty transparency log structure
The .emtpy files are required so that when the initial repository
is cloned, the log/leaves/pending, log/seq, and log/tile directories
will be created. The transparency tooling for adding entries assumes
these directories already exist and will not create them if they're
missing. Once we actually have actual log entries in the log, there
will be real files underneath these directories and this won't
be a problem any more.